Add build_instance hook in compute manager
The build_instance hook provides a hook close to the actual spawn
of an instance. It distinguishes itself from the create_instance
hook because, in the case of a successful build, the instance will
exist on the hypervisor when the build_instance post hook is
called, instead of simply just being scheduled as is the case with
the create_instance post hook. This is useful as some details may
not be availabe until after scheduling.

Change-Id: Ic12abf6315bbc6b010b8222ef296ec65cb4b4f95